removed shot_idx from trimmed file
shot_idx was causing a mismatch as it was holding the wrong index after the reserved shot was pulled out

Problem: The issue i believe is in that target_data already has the reserved shot taken out. Therefore, the indices are pulling from the wrong shot's frame count and that's why there's a discrepancy.